# openapi.yaml
openapi: 3.1.0
info:
  title: Process Documentation Agent API
  version: 1.0.0
  description: >
    MVP API for uploading video/audio/transcript inputs and generating
    PDD + SIPOC outputs with runtime provider selection (OpenAI/Google).
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/jobs:
    post:
      tags: [Jobs]
      summary: Create a processing job
      operationId: createJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - provider
              properties:
                provider:
                  $ref: "#/components/schemas/Provider"
                processing_profile:
                  $ref: "#/components/schemas/ProcessingProfile"
                context_notes:
                  type: string
                  maxLength: 2000
                video_file:
                  type: string
                  format: binary
                audio_file:
                  type: string
                  format: binary
                transcript_file:
                  type: string
                  format: binary
            encoding:
              video_file:
                contentType: video/mp4,video/quicktime,video/x-msvideo
              audio_file:
                contentType: audio/mpeg,audio/wav,audio/x-wav,audio/mp4
              transcript_file:
                contentType: text/plain,text/markdown,application/pdf
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Get job status and metadata
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetJobResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Jobs]
      summary: Delete job and artifacts
      operationId: deleteJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteJobResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/jobs/{job_id}/draft:
    get:
      tags: [Drafts]
      summary: Get generated draft for review
      operationId: getDraft
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Draft content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDraftResponse"
        "404":
          description: Job or draft not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [Drafts]
      summary: Save edited draft
      operationId: updateDraft
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDraftRequest"
      responses:
        "200":
          description: Draft saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateDraftResponse"
        "400":
          description: Invalid draft payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/jobs/{job_id}/finalize:
    post:
      tags: [Drafts]
      summary: Finalize reviewed draft and trigger export generation
      operationId: finalizeJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeRequest"
      responses:
        "202":
          description: Finalization accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizeResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Invalid state for finalization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/jobs/{job_id}/exports/{format}:
    get:
      tags: [Exports]
      summary: Download export artifact
      operationId: getExport
      parameters:
        - $ref: "#/components/parameters/JobId"
        - name: format
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ExportFormat"
      responses:
        "200":
          description: Export content
          content:
            text/markdown:
              schema:
                type: string
            application/json:
              schema:
                type: object
                additionalProperties: true
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Job or export not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Job not finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Provider:
      type: string
      enum: [openai, google]
    ProcessingProfile:
      type: string
      enum: [quality, balanced, low_cost]
      default: balanced
    JobStatus:
      type: string
      enum: [queued, processing, needs_review, completed, failed, expired]
    ExportFormat:
      type: string
      enum: [md, json, pdf]
    ApiEnvelope:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true
    ErrorObject:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: false
            data:
              type: "null"
            error:
              $ref: "#/components/schemas/ErrorObject"
    InputFileMeta:
      type: object
      required: [filename, content_type, size_bytes, storage_key]
      properties:
        filename:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
          minimum: 0
        storage_key:
          type: string
    InputManifest:
      type: object
      required: [video, audio, transcript]
      properties:
        video:
          anyOf:
            - $ref: "#/components/schemas/InputFileMeta"
            - type: "null"
        audio:
          anyOf:
            - $ref: "#/components/schemas/InputFileMeta"
            - type: "null"
        transcript:
          anyOf:
            - $ref: "#/components/schemas/InputFileMeta"
            - type: "null"
    LimitsApplied:
      type: object
      required:
        - max_file_size_mb
        - max_job_duration_seconds
        - max_provider_tokens
        - cost_target_band_usd_per_media_hour
      properties:
        max_file_size_mb:
          type: integer
          const: 500
        max_job_duration_seconds:
          type: integer
          minimum: 1
        max_provider_tokens:
          type: integer
          minimum: 1
        cost_target_band_usd_per_media_hour:
          type: object
          required: [min, max]
          properties:
            min:
              type: number
            max:
              type: number
    ModelPlan:
      type: object
      required:
        - provider
        - transcription_model
        - multimodal_model
        - generation_model
        - fallback_transcription
      properties:
        provider:
          $ref: "#/components/schemas/Provider"
        transcription_model:
          type: string
        multimodal_model:
          type: string
        generation_model:
          type: string
        fallback_transcription:
          type: object
          required: [provider, model]
          properties:
            provider:
              $ref: "#/components/schemas/Provider"
            model:
              type: string
    UsageCostEstimate:
      type: object
      required:
        - currency
        - estimated_total
        - estimated_per_media_hour
      properties:
        currency:
          type: string
          const: USD
        estimated_total:
          type: number
        estimated_per_media_hour:
          type: number
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        audio_seconds_processed:
          type: number
          minimum: 0
        video_seconds_processed:
          type: number
          minimum: 0
    ArtifactManifest:
      type: object
      required: [md, json, pdf]
      properties:
        md:
          anyOf:
            - type: string
            - type: "null"
        json:
          anyOf:
            - type: string
            - type: "null"
        pdf:
          anyOf:
            - type: string
            - type: "null"
    ProgressInfo:
      type: object
      required: [stage, percent]
      properties:
        stage:
          type: string
        percent:
          type: integer
          minimum: 0
          maximum: 100
    StepItem:
      type: object
      required:
        - step_no
        - title
        - actor
        - system
        - description
        - input
        - output
        - exception
      properties:
        step_no:
          type: integer
          minimum: 1
        title:
          type: string
        actor:
          type: string
        system:
          type: string
        description:
          type: string
        input:
          type: string
        output:
          type: string
        exception:
          type: string
    PDDDocument:
      type: object
      required:
        - purpose
        - scope
        - triggers
        - preconditions
        - steps
        - roles
        - systems
        - business_rules
        - exceptions
        - outputs
        - metrics
        - risks
      properties:
        purpose:
          type: string
        scope:
          type: string
        triggers:
          type: array
          items:
            type: string
        preconditions:
          type: array
          items:
            type: string
        steps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/StepItem"
        roles:
          type: array
          items:
            type: string
        systems:
          type: array
          items:
            type: string
        business_rules:
          type: array
          items:
            type: string
        exceptions:
          type: array
          items:
            type: string
        outputs:
          type: array
          items:
            type: string
        metrics:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
    SIPOCRow:
      type: object
      required: [supplier, input, process_step, output, customer]
      properties:
        supplier:
          type: string
        input:
          type: string
        process_step:
          type: string
        output:
          type: string
        customer:
          type: string
    ReviewFlag:
      type: object
      required: [type, path, message]
      properties:
        type:
          type: string
        path:
          type: string
        message:
          type: string
    ReviewNotes:
      type: object
      required: [quality_score, flags, assumptions]
      properties:
        quality_score:
          type: number
        flags:
          type: array
          items:
            $ref: "#/components/schemas/ReviewFlag"
        assumptions:
          type: array
          items:
            type: string
    JobRecord:
      type: object
      required:
        - id
        - status
        - provider
        - processing_profile
        - model_plan
        - input_manifest
        - limits_applied
        - usage_cost_estimate
        - artifacts
        - error_code
        - error_message
        - created_at
        - updated_at
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/JobStatus"
        provider:
          $ref: "#/components/schemas/Provider"
        processing_profile:
          $ref: "#/components/schemas/ProcessingProfile"
        model_plan:
          $ref: "#/components/schemas/ModelPlan"
        input_manifest:
          $ref: "#/components/schemas/InputManifest"
        limits_applied:
          $ref: "#/components/schemas/LimitsApplied"
        usage_cost_estimate:
          $ref: "#/components/schemas/UsageCostEstimate"
        artifacts:
          $ref: "#/components/schemas/ArtifactManifest"
        progress:
          $ref: "#/components/schemas/ProgressInfo"
        error_code:
          anyOf:
            - type: string
            - type: "null"
        error_message:
          anyOf:
            - type: string
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    CreateJobData:
      type: object
      required: [job_id, status, provider, processing_profile, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          const: queued
        provider:
          $ref: "#/components/schemas/Provider"
        processing_profile:
          $ref: "#/components/schemas/ProcessingProfile"
        created_at:
          type: string
          format: date-time
    CreateJobResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: "#/components/schemas/CreateJobData"
            error:
              type: "null"
    GetJobResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: "#/components/schemas/JobRecord"
            error:
              type: "null"
    DraftPayload:
      type: object
      required: [job_id, provider, model_plan, pdd, sipoc, review_notes, updated_at]
      properties:
        job_id:
          type: string
          format: uuid
        provider:
          $ref: "#/components/schemas/Provider"
        model_plan:
          $ref: "#/components/schemas/ModelPlan"
        pdd:
          $ref: "#/components/schemas/PDDDocument"
        sipoc:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SIPOCRow"
        review_notes:
          $ref: "#/components/schemas/ReviewNotes"
        updated_at:
          type: string
          format: date-time
    GetDraftResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: "#/components/schemas/DraftPayload"
            error:
              type: "null"
    UpdateDraftRequest:
      type: object
      required: [pdd, sipoc]
      properties:
        pdd:
          $ref: "#/components/schemas/PDDDocument"
        sipoc:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SIPOCRow"
        editor_notes:
          type: string
    UpdateDraftData:
      type: object
      required: [job_id, saved, updated_at]
      properties:
        job_id:
          type: string
          format: uuid
        saved:
          type: boolean
          const: true
        updated_at:
          type: string
          format: date-time
    UpdateDraftResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: "#/components/schemas/UpdateDraftData"
            error:
              type: "null"
    FinalizeRequest:
      type: object
      properties:
        finalize_note:
          type: string
    FinalizeData:
      type: object
      required: [job_id, status, next_stage]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [processing, completed]
        next_stage:
          type: string
    FinalizeResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: "#/components/schemas/FinalizeData"
            error:
              type: "null"
    DeleteJobData:
      type: object
      required: [job_id, deleted]
      properties:
        job_id:
          type: string
          format: uuid
        deleted:
          type: boolean
          const: true
    DeleteJobResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: "#/components/schemas/DeleteJobData"
            error:
              type: "null"
...